+++ test/flutter_local_ai_platform_interface_test.dart
@@ -48,6 +48,12 @@ class ImplementsFlutterLocalAiPlatform implements FlutterLocalAiPlatform {
   }
 
   @override
+  Stream<String> generateTextStream(
+      {required String prompt, GenerationConfig? config}) {
+    throw UnimplementedError();
+  }
+
+  @override
   Future<bool> openAICorePlayStore() {
     throw UnimplementedError();
   }